% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eive}
\alias{eive}
\title{Ecological Indicator Values for Europe}
\format{
A data frame with 47450 rows and 6 columns, the definitions of which are:
\describe{
\item{taxon_name}{The taxon name of the synonymous taxon concept associated with the recommended_taxon_name.}
\item{recommended_taxon_name}{The recommended taxon name for the taxon concept in the taxon_name column.}
\item{recommended_TVK}{The UKSI taxon version key associated with the recommended_taxon_name.}
\item{metric}{One of "optima" or "width".}
\item{variable}{The indicator value variable, one of: "L", "M", "N", "R", or "T".}
\item{value}{The source of the trait value.}
}
}
\usage{
eive
}
\description{
Ecological Indicator Values for Europe (EIVE) 1.0 (Dengler et al., ) values for taxa present in the UKSI.
}
\details{
\code{eive}
}
\references{
Dengler, J., Jansen, F., Chusova, O., Hüllbusch, E., Nobis, M. P., Meerbeek, K. V., Axmanová, I., Bruun, H. H., Chytrý, M., Guarino, R., Karrer, G., Moeys, K., Raus, T., Steinbauer, M. J., Tichý, L., Tyler, T., Batsatsashvili, K., Bita-Nicolae, C., Didukh, Y., … Gillet, F. (2023). Ecological Indicator Values for Europe (EIVE) 1.0. Vegetation Classification and Survey, 4, 7–29. https://doi.org/10.3897/VCS.98324
}
\keyword{datasets}
