% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylo_taxa_lookup}
\alias{phylo_taxa_lookup}
\title{Taxa lookup}
\format{
A data frame with 23477 rows and 4 columns, the definitions of which are:
\describe{
\item{TVK}{The UKSI taxon version key associated with the taxon_name.}
\item{taxon_name}{The recommended taxon name as present in the taxon_name column of \code{UKVegTB::taxonomic_backbone} and the recommended_taxon_name column of \code{UKVegTB::taxa_lookup}.}
\item{phylo_taxon_name}{The taxon names associated with the taxon_name column names as present in \code{UKVegTB::phylo_tree}.}
\item{phylo}{A boolean representing whether each taxon is present in the \code{UKVegTB::phylo_tree} object.}
}
}
\usage{
phylo_taxa_lookup
}
\description{
A lookup between the recommended taxa as present in \code{UKVegTB::taxonomic_backbone} and the taxon names present in the Newick \code{UKVegTB::phylo_tree} phylogenetic tree object.
At present 15246 taxa out of 23477 are present in the tree.
The phylo_taxon_name are produced by reducing varieties and subspecies to the species rank, and removing strata and ecological group suffixes.
}
\details{
\code{phylo_taxa_lookup}
}
\keyword{datasets}
