% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylo_tree}
\alias{phylo_tree}
\title{A phylogenetic tree for the UKVegTB taxa}
\format{
A string containing the phylogentic tree in the Newick format.
}
\usage{
phylo_tree
}
\description{
A phylogenetic tree in the Newick format, formed from taxa present in the phylo_taxon_name column in \code{UKVegTB::phylo_taxa_lookup}, where available.
}
\details{
\code{phylo_tree}
}
\keyword{datasets}
