% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_lookup}
\alias{taxa_lookup}
\title{Taxa lookup}
\format{
A data frame with 55622 rows and 13 columns, the definitions of which are:
\describe{
\item{informal_group}{The informal group of the taxon.}
\item{taxon_name}{The taxon name of the synonymous taxon concept associated with the recommended_taxon_name.}
\item{rank}{The taxonomic rank of the taxon.}
\item{qualifier}{The qualifier of the taxon.}
\item{authority}{The authority responsible for the description of the taxon.}
\item{full_name}{The full name of the taxon, formed from the combination of the taxon_name and qualifier.}
\item{TVK}{The UKSI taxon version key associated with the taxon_name.}
\item{recommended_taxon_name}{The recommended taxon name for the taxon concept in the taxon_name column.}
\item{recommended_rank}{The taxonomic rank of the recommended_taxon_name.}
\item{recommended_qualifier}{The qualifier of the recommended_taxon_name}
\item{recommended_authority}{The authority responsible for the description of the recommended_taxon_name.}
\item{recommended_full_name}{The full name of the recommended_taxon_name, formed from the combination of the recommended_taxon_name and recommended_qualifier.}
\item{recommended_TVK}{The UKSI taxon version key associated with the recommended_taxon_name.}
}
}
\usage{
taxa_lookup
}
\description{
A lookup between the recommended taxa as present in \code{UKVegTB::taxonomic_backbone} and associated synonymous taxon concepts present in the UKSI.
}
\details{
\code{taxa_lookup}
}
\keyword{datasets}
