% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomic_backbone}
\alias{taxonomic_backbone}
\title{Taxonomic Backbone}
\format{
A data frame with 17833 rows and 16 columns, the definitions of which are:
\describe{
\item{TVK}{}
\item{taxon_name}{}
\item{rank}{}
\item{qualifier}{}
\item{authority}{}
\item{full_name}{}
\item{informal_group}{}
\item{Domain}{}
\item{Kingdom}{}
\item{Division}{}
\item{Phylum}{}
\item{Class}{}
\item{Order}{}
\item{Family}{}
\item{Subfamily}{}
\item{Genus}{}
}
}
\usage{
taxonomic_backbone
}
\description{
The taxonomic backbone used in RMAVIS. Formed by selecting the recommended taxa from the following informal groups:
"alga", "chromist", "clubmoss", "conifer", "fern", "flowering plant", "ginkgo", "hornwort", "horsetail", "lichen", "liverwort", "moss", "quillwort", and "stonewort"
in version 20250703a of the UKSI, then retrieving the parent taxa.
}
\details{
\code{taxonomic_backbone}
}
\keyword{datasets}
