% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylo_taxa_lookup}
\alias{phylo_taxa_lookup}
\title{Taxa lookup}
\format{
A data frame with 17833 rows and 6 columns, the definitions of which are:
\describe{
\item{TVK}{The UKSI taxon version key associated with the taxon_name.}
\item{taxon_name}{The recommended taxon name as present in the taxon_name column of \code{UKVegTB::taxonomic_backbone} and the recommended_taxon_name column of \code{UKVegTB::taxa_lookup}.}
\item{search_name}{The taxon name used to search for OTL data, and as present in the \code{UKVegTB::phlyo_tree} object.}
\item{matched_name}{The OTL name matched to the search_name.}
\item{phylo}{A boolean representing whether each taxon is present in the \code{UKVegTB::phlyo_tree} object.}
\item{ottid}{The OTL id associated with the matched name.}
}
}
\usage{
phylo_taxa_lookup
}
\description{
A lookup between the recommended taxa as present in \code{UKVegTB::taxonomic_backbone} and the taxon names present in the Newick \code{UKVegTB::phylo_tree} phylogenetic tree object.
At present 11849 taxa out of 17833 are present in the tree.
The phylo_taxon_name are produced by reducing varieties and subspecies to the species rank, and removing strata and ecological group suffixes.
}
\details{
\code{phylo_taxa_lookup}
}
\keyword{datasets}
